Update movement.js
I added a new variable "start_transparency" (but it actually is a deceiving variable name and should be called "one_minus_start_transparency"). It takes a value between 0 and 1 and should be specified with the variables on lines 50-55 before you hit GO or NEXT for the simulation you want to capture frames from. It kind of adds the effect of motion blur and a value of 1 does nothing.